+++ Sources/CodexBar/PreferencesAboutPane.swift
@@ -103,10 +103,12 @@ struct AboutPane: View {
             guard !self.didLoadUpdaterState else { return }
             // Align Sparkle's flag with the persisted preference on first load.
             self.updater.automaticallyChecksForUpdates = self.autoUpdateEnabled
+            self.updater.automaticallyDownloadsUpdates = self.autoUpdateEnabled
             self.didLoadUpdaterState = true
         }
         .onChange(of: self.autoUpdateEnabled) { _, newValue in
             self.updater.automaticallyChecksForUpdates = newValue
+            self.updater.automaticallyDownloadsUpdates = newValue
         }
     }